Add checkStyle and internal role/permissions

+++ iam/src/main/java/ru/yandex/cloud/ml/platform/lzy/iam/grpc/LzySubjectService.java
@@ -7,11 +7,17 @@ import org.apache.logging.log4j.Logger;
 import ru.yandex.cloud.ml.platform.lzy.iam.authorization.AccessClient;
 import ru.yandex.cloud.ml.platform.lzy.iam.authorization.SubjectService;
 import ru.yandex.cloud.ml.platform.lzy.iam.authorization.exceptions.AuthException;
+import ru.yandex.cloud.ml.platform.lzy.iam.authorization.exceptions.AuthPermissionDeniedException;
+import ru.yandex.cloud.ml.platform.lzy.iam.grpc.context.AuthenticationContext;
+import ru.yandex.cloud.ml.platform.lzy.iam.resources.AuthPermission;
+import ru.yandex.cloud.ml.platform.lzy.iam.resources.impl.Root;
 import ru.yandex.cloud.ml.platform.lzy.iam.resources.subjects.Subject;
 import ru.yandex.cloud.ml.platform.lzy.iam.utils.GrpcConverter;
 import yandex.cloud.lzy.v1.IAM;
 import yandex.cloud.priv.lzy.v1.LSS;
 import yandex.cloud.priv.lzy.v1.LzySubjectServiceGrpc;
+
+import java.util.Objects;
 
 public class LzySubjectService extends LzySubjectServiceGrpc.LzySubjectServiceImplBase {
     public static final Logger LOG = LogManager.getLogger(LzySubjectService.class);
@@ -25,13 +31,15 @@ public class LzySubjectService extends LzySubjectServiceGrpc.LzySubjectServiceIm
     @Override
     public void createSubject(LSS.CreateSubjectRequest request, StreamObserver<IAM.Subject> responseObserver) {
         try {
-            Subject subject = subjectService.createSubject(
-                    request.getName(),
-                    request.getAuthProvider(),
-                    request.getProviderSubjectId()
-            );
-            responseObserver.onNext(GrpcConverter.from(subject));
-            responseObserver.onCompleted();
+            if (internalAccess()) {
+                Subject subject = subjectService.createSubject(
+                        request.getName(),
+                        request.getAuthProvider(),
+                        request.getProviderSubjectId()
+                );
+                responseObserver.onNext(GrpcConverter.from(subject));
+                responseObserver.onCompleted();
+            }
         } catch (AuthException e) {
             LOG.error("Auth exception::", e);
             responseObserver.onError(e.status().asException());
@@ -39,10 +47,14 @@ public class LzySubjectService extends LzySubjectServiceGrpc.LzySubjectServiceIm
     }
 
     @Override
-    public void removeSubject(LSS.RemoveSubjectRequest request, StreamObserver<LSS.RemoveSubjectResponse> responseObserver) {
+    public void removeSubject(
+            LSS.RemoveSubjectRequest request,
+            StreamObserver<LSS.RemoveSubjectResponse> responseObserver) {
         try {
-            subjectService.removeSubject(GrpcConverter.to(request.getSubject()));
-            responseObserver.onCompleted();
+            if (internalAccess()) {
+                subjectService.removeSubject(GrpcConverter.to(request.getSubject()));
+                responseObserver.onCompleted();
+            }
         } catch (AuthException e) {
             LOG.error("Auth exception::", e);
             responseObserver.onError(e.status().asException());
@@ -50,15 +62,19 @@ public class LzySubjectService extends LzySubjectServiceGrpc.LzySubjectServiceIm
     }
 
     @Override
-    public void addCredentials(LSS.AddCredentialsRequest request, StreamObserver<LSS.AddCredentialsResponse> responseObserver) {
+    public void addCredentials(
+            LSS.AddCredentialsRequest request,
+            StreamObserver<LSS.AddCredentialsResponse> responseObserver) {
         try {
-            subjectService.addCredentials(
-                    GrpcConverter.to(request.getSubject()),
-                    request.getCredentials().getName(),
-                    request.getCredentials().getCredentials(),
-                    request.getCredentials().getType()
-            );
-            responseObserver.onCompleted();
+            if (internalAccess()) {
+                subjectService.addCredentials(
+                        GrpcConverter.to(request.getSubject()),
+                        request.getCredentials().getName(),
+                        request.getCredentials().getCredentials(),
+                        request.getCredentials().getType()
+                );
+                responseObserver.onCompleted();
+            }
         } catch (AuthException e) {
             LOG.error("Auth exception::", e);
             responseObserver.onError(e.status().asException());
@@ -66,16 +82,31 @@ public class LzySubjectService extends LzySubjectServiceGrpc.LzySubjectServiceIm
     }
 
     @Override
-    public void removeCredentials(LSS.RemoveCredentialsRequest request, StreamObserver<LSS.RemoveCredentialsResponse> responseObserver) {
+    public void removeCredentials(
+            LSS.RemoveCredentialsRequest request,
+            StreamObserver<LSS.RemoveCredentialsResponse> responseObserver) {
         try {
-            subjectService.removeCredentials(
-                    GrpcConverter.to(request.getSubject()),
-                    request.getCredentialsName()
-            );
-            responseObserver.onCompleted();
+            if (internalAccess()) {
+                subjectService.removeCredentials(
+                        GrpcConverter.to(request.getSubject()),
+                        request.getCredentialsName()
+                );
+                responseObserver.onCompleted();
+            }
         } catch (AuthException e) {
             LOG.error("Auth exception::", e);
             responseObserver.onError(e.status().asException());
         }
+    }
+
+    private boolean internalAccess() {
+        var currentSubject = Objects.requireNonNull(AuthenticationContext.current()).getSubject();
+        if (!accessClient.hasResourcePermission(currentSubject,
+                new Root().resourceId(),
+                AuthPermission.INTERNAL_AUTHORIZE)) {
+            LOG.error("Not INTERNAL user::{} try to create subjects", currentSubject.id());
+            throw new AuthPermissionDeniedException("");
+        }
+        return true;
     }
 }
